// src/models/public/PlaceReportQuantities.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { PlaceReportsPlaceReportId } from './PlaceReports';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.place_report_quantities */
export type PlaceReportQuantitiesPlaceReportQuantityId = string & { __brand: 'public.place_report_quantities' };

/**
 * Represents the table public.place_report_quantities
 * Quantities of place reports
 */
export default interface PlaceReportQuantities {
  place_report_quantity_id: PlaceReportQuantitiesPlaceReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  place_report_id: PlaceReportsPlaceReportId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer: number | null;

  /** Used for numeric quantities */
  quantity_numeric: number | null;

  /** Used for text quantities */
  quantity_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.place_report_quantities
 * Quantities of place reports
 */
export interface PlaceReportQuantitiesInitializer {
  /** Default value: uuid_generate_v7() */
  place_report_quantity_id?: PlaceReportQuantitiesPlaceReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_report_id?: PlaceReportsPlaceReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer?: number | null;

  /** Used for numeric quantities */
  quantity_numeric?: number | null;

  /** Used for text quantities */
  quantity_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.place_report_quantities
 * Quantities of place reports
 */
export interface PlaceReportQuantitiesMutator {
  place_report_quantity_id?: PlaceReportQuantitiesPlaceReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_report_id?: PlaceReportsPlaceReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer?: number | null;

  /** Used for numeric quantities */
  quantity_numeric?: number | null;

  /** Used for text quantities */
  quantity_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}